// src/types/input-type.tsx
export type SubmissionType =
  | "Suggestion"
  | "Concern"
  | "Idea"
  | "Praise"
  | "Other (please specify)";

export type TopicArea =
  | "Academics"
  | "Organization/Leadership"
  | "Events or Activities"
  | "Student Well-being"
  | "Facilities or Resources"
  | "Other (please specify)";

export type Priority = "Urgent" | "Needs follow-up" | "For awareness only";